Implemented a workaround for Three20's table cell image layout bug (thanks RoBak42 for the patch)
There is a bug in Three20's table cell image logic w.r.t.
Three20's image cache. By applying a TTImageStyle to the table item,
we can override the layout logic to force the image to always be a fixed size.
Otherwise, the image will sometimes overflow the bounds of the cell.

This commit fixes the following issue:
http://github.com/klazuka/TTRemoteExamples/issues#issue/1